+++ src/test/groovy/com/github/gradle/node/npm/task/NpmTaskTest.groovy
@@ -16,7 +16,7 @@ class NpmTaskTest extends AbstractTaskTest {
     def "exec npm task"() {
         given:
         props.setProperty('os.name', 'Linux')
-        nodeExtension.computedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
+        nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
 
         def task = project.tasks.create('simple', NpmTask)
         mockPlatformHelper(task)
@@ -40,7 +40,7 @@ class NpmTaskTest extends AbstractTaskTest {
     def "exec npm task (windows)"() {
         given:
         props.setProperty('os.name', 'Windows')
-        nodeExtension.computedPlatform.set(PlatformHelperKt.parsePlatform("Windows", "x86_64", {}))
+        nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Windows", "x86_64", {}))
 
         def task = project.tasks.create('simple', NpmTask)
         mockPlatformHelper(task)
@@ -64,7 +64,7 @@ class NpmTaskTest extends AbstractTaskTest {
     def "exec npm task (download)"() {
         given:
         props.setProperty('os.name', 'Linux')
-        nodeExtension.computedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
+        nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
         nodeExtension.download.set(true)
         def nodeDir = projectDir.toPath().resolve(".gradle").resolve("nodejs")
                 .resolve("node-v${DEFAULT_NODE_VERSION}-linux-x64")
@@ -95,7 +95,7 @@ class NpmTaskTest extends AbstractTaskTest {
     def "exec npm task (download) with configured proxy"() {
         given:
         props.setProperty('os.name', 'Linux')
-        nodeExtension.computedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
+        nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
         nodeExtension.download.set(true)
         def nodeDir = projectDir.toPath().resolve(".gradle").resolve("nodejs")
                 .resolve("node-v${DEFAULT_NODE_VERSION}-linux-x64")
@@ -129,7 +129,7 @@ class NpmTaskTest extends AbstractTaskTest {
     def "exec npm task with configured proxy"() {
         given:
         props.setProperty('os.name', 'Linux')
-        nodeExtension.computedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
+        nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
         GradleProxyHelper.setHttpsProxyHost("my-super-proxy.net")
         GradleProxyHelper.setHttpsProxyPort(11235)
 
@@ -151,7 +151,7 @@ class NpmTaskTest extends AbstractTaskTest {
     def "exec npm task with configured proxy but disabled"() {
         given:
         props.setProperty('os.name', 'Linux')
-        nodeExtension.computedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
+        nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
         GradleProxyHelper.setHttpsProxyHost("my-super-proxy.net")
         GradleProxyHelper.setHttpsProxyPort(11235)
         nodeExtension.nodeProxySettings.set(ProxySettings.OFF)