+++ src/test/groovy/com/github/gradle/node/yarn/task/YarnTaskTest.groovy
@@ -33,7 +33,7 @@ class YarnTaskTest extends AbstractTaskTest {
     def "exec yarn task (download)"() {
         given:
         props.setProperty('os.name', 'Linux')
-        nodeExtension.computedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
+        nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
         nodeExtension.download.set(true)
 
         def task = project.tasks.create('simple', YarnTask)
@@ -57,7 +57,7 @@ class YarnTaskTest extends AbstractTaskTest {
     def "exec yarn task (download) with configured proxy"() {
         given:
         props.setProperty('os.name', 'Linux')
-        nodeExtension.computedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
+        nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
         nodeExtension.download.set(true)
         GradleProxyHelper.setHttpsProxyHost("1.2.3.4")
         GradleProxyHelper.setHttpsProxyPort(443)
@@ -86,7 +86,7 @@ class YarnTaskTest extends AbstractTaskTest {
     def "exec yarn task (download) with configured proxy but disabled"() {
         given:
         props.setProperty('os.name', 'Linux')
-        nodeExtension.computedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
+        nodeExtension.resolvedPlatform.set(PlatformHelperKt.parsePlatform("Linux", "x86_64", {}))
         nodeExtension.download.set(true)
         GradleProxyHelper.setHttpsProxyHost("1.2.3.4")
         GradleProxyHelper.setHttpsProxyPort(443)